music/serializers.py
from rest_framework import serializers
from .models import Song


class SongSerializer(serializers.ModelSerializer):
    class Meta:
        model = Song
        fields = ["id", "name", "slug", "cover_image", "artist", "release_date"]